requirements.txt
langchain>=0.1.14,<0.2.0
langchain-community>=0.0.21
langchain-openai>=0.0.8
langchain-text-splitters>=0.0.2
faiss-cpu>=1.7.4
pypdf>=4.2.0
python-dotenv>=1.0.0
streamlit>=1.30.0
tiktoken>=0.6.0
sentence-transformers>=2.3.0
huggingface-hub>=0.20.3
